+++ CHANGES.md
@@ -2,9 +2,9 @@
 
 ## 1.0.7 / NEXT RELEASE
 
-* introduce ServiceCall.invokeAll service endpoints 
-* introduce metrics module collecting KPIs
-* add service request / response / error rates KPIs 
+* Introduce ServiceCall.invokeAll service endpoints 
+* Introduce metrics module and codahle provider for collecting KPIs
+* Metrics added to ServiceCall request / response / error rates KPIs 
 
 ## 1.0.6 / 2017-09-20